Factor out SoftDeletes into module
