+++ src/Schema/AST/ASTBuilder.php
@@ -95,7 +95,6 @@ class ASTBuilder
         $this->addPaginationInfoTypes();
         $this->addNodeSupport();
         $this->addOrderByTypes();
-        $this->addTrashedEnum();
 
         // Listeners may manipulate the DocumentAST that is passed by reference
         // into the ManipulateAST event. This can be useful for extensions
@@ -358,26 +357,6 @@ GRAPHQL
                 input OrderByClause {
                    field: String!
                    order: SortOrder!
-                }
-            ')
-        );
-    }
-
-    /**
-     * Add Trashed enum to filter soft deleted models.
-     *
-     * @see \Nuwave\Lighthouse\Schema\Directives\TrashedDirective
-     *
-     * @return void
-     */
-    protected function addTrashedEnum(): void
-    {
-        $this->documentAST->setTypeDefinition(
-            PartialParser::enumTypeDefinition('
-                enum Trashed {
-                    ONLY @enum(value: "only")
-                    WITH @enum(value: "with")
-                    WITHOUT @enum(value: "without")
                 }
             ')
         );
